# docs/architecture/contracts/rest/openapi.yaml
openapi: 3.1.0
info:
  title: Mintkey Admin REST API
  version: "0.1.0-experimental"
  summary: Operator-facing control-plane HTTP surface for Mintkey.
  description: |
    Mintkey is a credential broker for AI agents. This API is consumed by:
      - The Admin Console (AdminJS) for sensitive operations.
      - Operator-owned tooling (CI, scripts) using API tokens.
      - Internal subscribers calling `/v1/changes` for reconciliation
        (per ADR-0010 change-channel model).

    Stability tier: **experimental**. Breaking changes will bump the major
    `info.version`; surface stability is per-resource and called out in each
    operation `description` where it differs from the document-level tier.

    ## ID conventions

    All identifiers are **ULIDs** (Crockford base32, 26 characters) prefixed
    by their resource type. Tenants are addressed at the wire by their prefixed ULID `tenant_…` on
    every entity for RLS purposes (ADR-0008).

    | Prefix          | Resource                          |
    | --------------- | --------------------------------- |
    | `tenant_…`      | Tenant                            |
    | `operator_…`    | Operator                          |
    | `agent_…`       | Agent                             |
    | `svc_…`         | Service                           |
    | `cred_…`        | Credential (key-version envelope) |
    | `perm_…`        | PermissionGrant                   |
    | `audit_…`       | AuditEvent                        |
...

...
servers:
...
security:
  - OperatorSession: []
  - OperatorBearer: []
paths:
...
  /v1/tenants/{tenant_id}/agents/{agent_id}/api-keys:
    parameters:
      - $ref: "#/components/parameters/TenantId"
      - $ref: "#/components/parameters/AgentId"
    post:
      tags: [APIKeys]
      operationId: createServiceApiKey
      summary: Create a service API key for an agent.
      description: |
        Generates a long-lived opaque API key (`mk_svckey_…`) scoped to a
        specific service. The plaintext key is returned **once** in the 201
        response and is never stored — store it immediately.

        `allowed_actions` must be a subset of the agent's existing permission
        grants for the service. Operator policies (`AdminSettings.api_key`)
        may impose additional constraints (require_expiry, max_expiry_days,
        require_ip_allowlist).

        Req 1.1–1.6, ADR-0018 §4.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceApiKeyCreate"
      responses:
        "201":
          description: Key created. Plaintext shown once.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceApiKeyCreated"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/BadRequest"
    get:
      tags: [APIKeys]
      operationId: listServiceApiKeys
      summary: List service API keys for an agent.
      description: |
        Returns all keys (active, expired, revoked) for the agent. Filtered
        by `service_id` and/or `status` if query params provided.
        No plaintext is returned. Req 8.2, 8.6.
      parameters:
        - in: query
          name: service_id
          required: false
          schema:
            type: string
            pattern: "^svc_[0-9A-HJKMNP-TV-Z]{26}$"
          description: Filter to a specific service.
        - in: query
          name: status
          required: false
          schema:
            type: string
            enum: [active, expired, revoked]
          description: Filter by key status.
        - in: query
          name: q
          required: false
          schema:
            type: string
            maxLength: 200
          description: >
            Case-insensitive substring/prefix search on key_fingerprint.
            Useful to locate a key by its first few fingerprint characters.
        - $ref: "#/components/parameters/AfterCursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Paginated key list.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/ServiceApiKey"
                  next_cursor:
                    type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/tenants/{tenant_id}/agents/{agent_id}/api-keys/{api_key_id}:
    parameters:
      - $ref: "#/components/parameters/TenantId"
      - $ref: "#/components/parameters/AgentId"
      - in: path
        name: api_key_id
        required: true
        schema:
          type: string
          pattern: "^svckey_[0-9A-HJKMNP-TV-Z]{26}$"
        description: Service API key identifier.
    get:
      tags: [APIKeys]
      operationId: getServiceApiKey
      summary: Get a single service API key.
      description: |
        Returns key metadata. No plaintext. Req 8.3, 8.6.
      responses:
        "200":
          description: Key metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceApiKey"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/tenants/{tenant_id}/agents/{agent_id}/api-keys/{api_key_id}/revoke:
    parameters:
      - $ref: "#/components/parameters/TenantId"
      - $ref: "#/components/parameters/AgentId"
      - in: path
        name: api_key_id
        required: true
        schema:
          type: string
          pattern: "^svckey_[0-9A-HJKMNP-TV-Z]{26}$"
        description: Service API key identifier.
    post:
      tags: [APIKeys]
      operationId: revokeServiceApiKey
      summary: Revoke a service API key.
      description: |
        Marks the key revoked. Idempotent — already-revoked returns 200.
        Emits `api_key.revoked` audit event and publishes `pg_notify` so the
        proxy evicts the cached resolution immediately. Req 4.1, 4.5, 8.4.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                reason:
                  type: string
                  maxLength: 512
      responses:
        "200":
          description: Revoked (or was already revoked).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceApiKey"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/tenants/{tenant_id}/agents/{agent_id}/api-keys/{api_key_id}/rotate:
    parameters:
      - $ref: "#/components/parameters/TenantId"
      - $ref: "#/components/parameters/AgentId"
      - in: path
        name: api_key_id
        required: true
        schema:
          type: string
          pattern: "^svckey_[0-9A-HJKMNP-TV-Z]{26}$"
        description: Service API key identifier.
    post:
      tags: [APIKeys]
      operationId: rotateServiceApiKey
      summary: Rotate a service API key.
      description: |
        Creates a new key with the same binding (agent, service, allowed_actions,
        constraints). Does NOT revoke the old key — the operator controls
        overlap. Returns the new key's plaintext once. Emits `api_key.rotated`
        audit event. Req 5.1, 5.2, 8.5.
      responses:
        "201":
          description: New key created. Plaintext shown once.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceApiKeyCreated"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  ###########################################################################
  # Audit
  ###########################################################################
...
components:
...
  securitySchemes:
    OperatorSession:
      type: apiKey
      in: cookie
      name: mintkey_session
      description: |
        HttpOnly Secure SameSite=Strict session cookie issued by
        `/v1/auth/internal-login` or `/v1/auth/callback`.
    OperatorBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Operator-owned API token (PAT) for machine-to-machine calls. Issued
        from the Admin Console; not exposed in this OpenAPI document. Sent
        as `Authorization: Bearer …`.
    AgentApiKey:
      type: http
      scheme: bearer
      bearerFormat: mintkey-agent-api-key
      description: |
        Long-lived Agent API Key (`mk_live_…`). Used by the few agent-facing
        REST endpoints; most agent traffic flows through MCP instead.

    AdminUiSignedRequest:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Ed25519 JWT signed by the AdminJS service. Required on every endpoint
        AdminJS calls on behalf of an operator. Carries claims:
          iss: "mintkey/admin-ui"
          sub: <operator_id>
          tnt: <tenant_id>
          aud: "mintkey/admin-api"
          iat, exp (60s TTL)
          jti (UUID; replay-checked against admin_request_jti per ADR-0016.1)
        FastAPI verifies the signature against AdminJS's public key fetched from
        the Vault Adapter under credential type 'admin_ui_signing_key'.
        Per ADR-0014.6 + ADR-0017.1.
    ServiceIdentity:
...
  parameters:
...
    TenantId:
      in: path
      name: tenant_id
      required: true
      schema:
        type: string
        pattern: "^tenant_[0-9A-HJKMNP-TV-Z]{26}$"
      description: ULID-prefixed tenant identifier.
...
    AgentId:
      in: path
      name: agent_id
      required: true
      schema:
        type: string
        pattern: "^agent_[0-9A-HJKMNP-TV-Z]{26}$"
      description: ULID-prefixed agent identifier.
    AfterCursor:
      in: query
      name: after
      required: false
      schema:
        type: string
      description: |
        Cursor — the `id` of the last item from the previous page. Exclusive.
...
  responses:
...
    Unauthorized:
      description: No or invalid credentials.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: "about:blank"
            title: "Unauthorized"
            status: 401
            detail: "Session cookie missing or invalid."
            instance: "/v1/auth/whoami"
            "mintkey:code": "unauthenticated"
            "mintkey:trace_id": "0af7651916cd43dd8448eb211c80319c"
    Forbidden:
      description: Authenticated but not allowed.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: "about:blank"
            title: "Forbidden"
            status: 403
            detail: "Operator role 'Auditor' cannot create services."
            instance: "/v1/tenants/tenant_01HX5J9F8V8H8V0CG3F2Y5J6M9/services"
            "mintkey:code": "permission_denied"
            "mintkey:trace_id": "0af7651916cd43dd8448eb211c80319c"
    NotFound:
      description: Resource does not exist or is not visible to the caller.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: "about:blank"
            title: "Not Found"
            status: 404
            detail: "No tenant with id 'tenant_01HX5J9F8V8H8V0CG3F2Y5J6M9'."
            instance: "/v1/tenants/tenant_01HX5J9F8V8H8V0CG3F2Y5J6M9"
            "mintkey:code": "tenant_not_found"
            "mintkey:trace_id": "0af7651916cd43dd8448eb211c80319c"
...
  schemas:
...
    ServiceApiKeyCreate:
      type: object
      additionalProperties: false
      description: |
        Request body for creating a service API key. `allowed_actions` must
        be a non-empty subset of the agent's permission grants for
        `service_id`. Req 1.1, 8.1.
      required: [service_id, allowed_actions]
      properties:
        service_id:
          type: string
          pattern: "^svc_[0-9A-HJKMNP-TV-Z]{26}$"
          description: Service the key is bound to.
        allowed_actions:
          type: array
          minItems: 1
          items:
            type: string
          description: Actions the key may perform (subset of agent grants).
        expires_at:
          type: ["string", "null"]
          format: date-time
          description: |
            Optional expiry (RFC 3339 UTC). Must be in the future and
            honour `AdminSettings.api_key.max_expiry_days`.
        constraints:
          $ref: "#/components/schemas/Constraints"
...
    ServiceApiKeyCreated:
      allOf:
        - $ref: "#/components/schemas/ServiceApiKey"
        - type: object
          required: [plaintext_key]
          properties:
            plaintext_key:
              type: string
              pattern: "^mk_svckey_"
              x-mintkey-sensitive: true
              description: |
                Full plaintext key. Shown **once** at creation or rotation.
                Never stored — present only in this 201 response. Req 1.2.
...